// src/app/shared/models/message-parameter.ts
import { SweetAlertIcon } from 'sweetalert2';

export interface MessageParameter {
  title: string;
  icon: SweetAlertIcon;
  position?:
    | 'top'
    | 'top-start'
    | 'top-end'
    | 'top-left'
    | 'top-right'
    | 'center'
    | 'center-start'
    | 'center-end'
    | 'center-left'
    | 'center-right'
    | 'bottom'
    | 'bottom-start'
    | 'bottom-end'
    | 'bottom-left'
    | 'bottom-right';
}